remove everything but the s3-to-postgres stuff
I decided to remove everything except the copying down from s3.
We're not going to be using that stuff again, and if we did, we'd
recreate it some other way anyway. This might be slightly rough
if this is used externally, but I'm not sure it is and it's also rough
to have code half-supported lying around.