Add support for 'stream' granularity
For non-snapshot data it makes more sense to just the latest data into
the db as we go. This means each s3 payload just has the new data so we
don't need to worry about truncating or anything like that.

This does mean that these jobs are no longer idempotent which is
unfortunate... It makes me wonder whether we should merge record-keeper
and s3-to-redshift since breaking apart figuring out what to write and
actually writing it limits what we can do safely.